add func genChSet
and refactoring structure

+++ firmware/BTGenerator/generator.h
@@ -18,10 +18,10 @@ struct GenChannel
 {
   FormChannel form;
 
-  //Обработанные значения амплитуды, частоты и фазы (0...100, 0...1000.00, 0...360)
-  uint16_t amp;
+  //Амплитуда, частота и фаза (0...100.00, 0...1000.00, 0...360.00)
+  float amp;
   float freq;
-  uint16_t phase;
+  float phase;
 
   //Предварительные вычисленные удобные значения
   //для ускорения работы прерываний
@@ -37,6 +37,6 @@ extern GenChannel channels[CHANNELS_COUNT];
 
 void genInit(); //Инициализация каналов нулями
 void genSync(); //Функция принудительной синхронизации всех каналов
-void genChSet(uint8_t channel, FormChannel form, uint16_t amp, float freq, uint16_t phase);//Функция установки параметров канала
+void genChSet(uint8_t channel, FormChannel form, float amp, float freq, float phase);//Функция установки параметров канала
 
 #endif
